+++ packages/eyes-sdk-core/lib/utils/ConfigUtils.js
@@ -7,7 +7,7 @@ const Logger = require('../logging/Logger')
 function getConfig({
   configParams = [],
   configPath,
-  logger = new Logger(process.env.APPLITOOLS_SHOW_LOGS),
+  logger = new Logger(!!process.env.APPLITOOLS_SHOW_LOGS),
 } = {}) {
   const explorer = cosmiconfigSync('applitools', {
     searchPlaces: ['applitools.config.js', 'package.json', 'eyes.config.js', 'eyes.json'],