+++ packages/eyes-storybook/test/it/processResults.test.js
@@ -248,6 +248,7 @@ describe('processResults', () => {
     expect(expectedOutput).to.eql(outputStr);
     expect(exitCode).to.eql(1);
   });
+
   it('passes errors to the formatter correctly', async () => {
     const results = [
       {
@@ -260,5 +261,19 @@ describe('processResults', () => {
     expect(storedResults.length).to.eql(1);
     expect(storedResults[0].getName()).to.eql('My Component | Button1');
     expect(storedResults[0].error).to.eql(results[0].resultsOrErr[0]);
+  });
+
+  it('passes errors at the story level (not the rendering in vgc) to the formatter correctly', async () => {
+    const results = [
+      {
+        title: 'My Component | Button1',
+        resultsOrErr: new Error('some error message thrown e.g. inside getStoryData'),
+      },
+    ];
+    const {formatter} = processResults({results, totalTime: 10000, concurrency: 1});
+    const storedResults = formatter.getResultsList();
+    expect(storedResults.length).to.eql(1);
+    expect(storedResults[0].getName()).to.eql('My Component | Button1');
+    expect(storedResults[0].error).to.eql(results[0].resultsOrErr);
   });
 });