+++ packages/sdk-shared/src/test-setup.js
@@ -63,7 +63,7 @@ const BROWSERS = {
     capabilities: {
       name: 'Edge 18',
       browserName: 'MicrosoftEdge',
-      browserVersion: '18',
+      version: '18.17763',
       platformName: 'Windows 10',
       screenResolution: '1920x1080',
       ...SAUCE_CREDENTIALS,
@@ -110,11 +110,8 @@ const BROWSERS = {
   firefox: {
     capabilities: {
       browserName: 'firefox',
-      seleniumVersion: '3.141.59',
-      ...SAUCE_CREDENTIALS,
     },
-    url: SAUCE_SERVER_URL,
-    sauce: true,
+    url: 'http://localhost:4445/wd/hub',
   },
   chrome: {
     capabilities: {
@@ -150,7 +147,13 @@ function Env(
 const batchName = process.env.APPLITOOLS_BATCH_NAME || 'JS Coverage Tests'
 const batch = typeof BatchInfo === 'undefined' ? batchName : new BatchInfo(batchName)
 
-function getEyes({isVisualGrid, isCssStitching, configuration, branchName = 'master'} = {}) {
+function getEyes({
+  isVisualGrid,
+  isCssStitching,
+  configuration,
+  branchName = 'master',
+  showLogs,
+} = {}) {
   const eyes = new Eyes(isVisualGrid ? new VisualGridRunner(10) : undefined)
   const conf = Object.assign(
     {
@@ -167,7 +170,7 @@ function getEyes({isVisualGrid, isCssStitching, configuration, branchName = 'mas
   )
   eyes.setConfiguration(new Configuration(conf))
 
-  if (process.env.APPLITOOLS_SHOW_LOGS) {
+  if (process.env.APPLITOOLS_SHOW_LOGS || showLogs) {
     eyes.setLogHandler(new ConsoleLogHandler(true))
   }