+++ packages/sdk-shared/test/check-for-docker-hostname.spec.js
@@ -1,0 +1,10 @@
+const assert = require('assert')
+const checkForDockerHostname = require('../src/check-for-docker-hostname')
+
+describe('checkForDockerHostname', () => {
+  it('does not run on Linux', () => {
+    return assert.doesNotThrow(async () => {
+      await checkForDockerHostname({platform: 'linux'})
+    })
+  })
+})