+++ packages/visual-grid-client/test/unit/sdk/createRenderRequests.test.js
@@ -6,7 +6,7 @@ const createRenderRequests = require('../../../src/sdk/createRenderRequests')
 const createRGridDom = require('../../../src/sdk/createRGridDom')
 
 describe('createRenderRequests', () => {
-  let url, renderInfo, dom, rGridDom
+  let url, renderInfo, dom, domObj, resources, resourcesObj
 
   beforeEach(() => {
     url = 'url'
@@ -15,12 +15,14 @@ describe('createRenderRequests', () => {
       getStitchingServiceUrl: () => 'stitchingServiceUrl',
     }
     const cdt = 'cdt'
-    dom = {
+    domObj = {
       contentType: 'x-applitools-html/cdt',
       hash: getSha256Hash(JSON.stringify({resources: {}, domNodes: cdt})),
       hashFormat: 'sha256',
     }
-    rGridDom = createRGridDom({resources: {}, cdt})
+    dom = createRGridDom({resources: {}, cdt})
+    resources = []
+    resourcesObj = {}
   })
 
   it('works', () => {
@@ -29,6 +31,7 @@ describe('createRenderRequests', () => {
     const url = 'url'
     const cdt = 'cdt'
     const resources = [r1, r2]
+    const dom = createRGridDom({resources: {['url1']: r1, ['url2']: r2}, cdt})
     const browsers = [
       {width: 1, height: 2, name: 'b1'},
       {width: 3, height: 4, name: 'b2'},
@@ -41,8 +44,7 @@ describe('createRenderRequests', () => {
 
     const renderRequests = createRenderRequests({
       url,
-      resources,
-      dom: createRGridDom({resources: {['url1']: r1, ['url2']: r2}, cdt}),
+      pages: Array(browsers.length).fill({rGridDom: dom, allResources: resources}),
       browsers,
       renderInfo,
       sizeMode,
@@ -55,7 +57,7 @@ describe('createRenderRequests', () => {
     })
 
     const resourcesObj = {url1: 'hash1', url2: 'hash2'}
-    const dom = {
+    const domObj = {
       contentType: 'x-applitools-html/cdt',
       hash: getSha256Hash(JSON.stringify({resources: resourcesObj, domNodes: 'cdt'})),
       hashFormat: 'sha256',
@@ -66,7 +68,7 @@ describe('createRenderRequests', () => {
         webhook: 'resultsUrl',
         stitchingService: 'stitchingServiceUrl',
         url,
-        dom,
+        dom: domObj,
         resources: resourcesObj,
         browser: {name: 'b1'},
         scriptHooks,
@@ -83,7 +85,7 @@ describe('createRenderRequests', () => {
         webhook: 'resultsUrl',
         stitchingService: 'stitchingServiceUrl',
         url,
-        dom,
+        dom: domObj,
         resources: resourcesObj,
         browser: {name: 'b2'},
         scriptHooks,
@@ -105,8 +107,7 @@ describe('createRenderRequests', () => {
     const browsers = [{deviceName, screenOrientation}]
     const renderRequests = createRenderRequests({
       url,
-      resources: [],
-      dom: rGridDom,
+      pages: [{rGridDom: dom, allResources: resources}],
       browsers,
       renderInfo,
       noOffsetSelectors: [],
@@ -118,8 +119,8 @@ describe('createRenderRequests', () => {
         webhook: 'resultsUrl',
         stitchingService: 'stitchingServiceUrl',
         url,
-        dom,
-        resources: {},
+        dom: domObj,
+        resources: resourcesObj,
         renderInfo: {
           emulationInfo: {deviceName, screenOrientation},
           height: undefined,
@@ -140,8 +141,7 @@ describe('createRenderRequests', () => {
     }
     const renderRequests = createRenderRequests({
       url,
-      resources: [],
-      dom: rGridDom,
+      pages: [{rGridDom: dom, allResources: resources}],
       browsers,
       renderInfo,
       noOffsetSelectors: [],
@@ -153,8 +153,8 @@ describe('createRenderRequests', () => {
         webhook: 'resultsUrl',
         stitchingService: 'stitchingServiceUrl',
         url,
-        dom,
-        resources: {},
+        dom: domObj,
+        resources: resourcesObj,
         renderInfo: {
           emulationInfo: {
             width: 1,
@@ -187,8 +187,7 @@ describe('createRenderRequests', () => {
     const floating = [{some: 'thing'}, {type: 'css', selector: 'sel'}]
     const renderRequests = createRenderRequests({
       url,
-      resources: [],
-      dom: rGridDom,
+      pages: [{rGridDom: dom, allResources: resources}],
       browsers,
       renderInfo,
       noOffsetSelectors: [ignore, layout, strict, content, accessibility],
@@ -200,8 +199,8 @@ describe('createRenderRequests', () => {
         webhook: 'resultsUrl',
         stitchingService: 'stitchingServiceUrl',
         url,
-        dom,
-        resources: {},
+        dom: domObj,
+        resources: resourcesObj,
         renderInfo: {
           height: 2,
           width: 1,
@@ -233,8 +232,7 @@ describe('createRenderRequests', () => {
     const browsers = [{iosDeviceInfo}]
     const renderRequests = createRenderRequests({
       url,
-      resources: [],
-      dom: rGridDom,
+      pages: [{rGridDom: dom, allResources: resources}],
       browsers,
       renderInfo,
     })
@@ -244,8 +242,8 @@ describe('createRenderRequests', () => {
         webhook: 'resultsUrl',
         stitchingService: 'stitchingServiceUrl',
         url,
-        dom,
-        resources: {},
+        dom: domObj,
+        resources: resourcesObj,
         browser: {name: 'safari'},
         platform: {name: 'ios'},
         renderInfo: {
@@ -258,5 +256,47 @@ describe('createRenderRequests', () => {
         },
       },
     ])
+  })
+
+  it('handles multiple dom snapshots', () => {
+    const browsers = [
+      {width: 1, height: 1, name: '1'},
+      {width: 2, height: 2, name: '2'},
+      {width: 3, height: 3, name: '1'},
+    ]
+    const dom1 = createRGridDom({resources: {}, cdt: 'cdt1'})
+    const dom2 = createRGridDom({resources: {}, cdt: 'cdt2'})
+    const renderRequests = createRenderRequests({
+      url,
+      pages: [
+        {rGridDom: dom1, allResources: resources},
+        {rGridDom: dom2, allResources: resources},
+        {rGridDom: dom1, allResources: resources},
+      ],
+      browsers,
+      renderInfo,
+    })
+
+    expect(renderRequests.map(r => r.toJSON())).to.eql(
+      browsers.map(browser => ({
+        webhook: 'resultsUrl',
+        stitchingService: 'stitchingServiceUrl',
+        url,
+        dom: {
+          contentType: 'x-applitools-html/cdt',
+          hash: getSha256Hash(JSON.stringify({resources: {}, domNodes: `cdt${browser.name}`})),
+          hashFormat: 'sha256',
+        },
+        resources: resourcesObj,
+        browser: {name: browser.name},
+        renderInfo: {
+          width: browser.width,
+          height: browser.height,
+          region: undefined,
+          selector: undefined,
+          sizeMode: undefined,
+        },
+      })),
+    )
   })
 })