+++ packages/eyes-playwright/CHANGELOG.md
@@ -8,3 +8,5 @@
 - updated to @applitools/eyes-sdk-core@12.0.1 (from 11.5.0)
 - updated to @applitools/visual-grid-client@14.6.3 (from 14.6.0)
 - updated to @applitools/visual-grid-client@14.6.4 (from 14.6.3)
+- updated to @applitools/eyes-sdk-core@12.2.0 (from 12.0.1)
+- updated to @applitools/visual-grid-client@15.0.0 (from 14.6.4)