+++ packages/eyes-protractor/CHANGELOG.md
@@ -2,7 +2,17 @@
 
 ## Unreleased
 
+
+## 1.8.0 - 2020/9/3
+
+- updated to @applitools/eyes-sdk-core@12.2.3 (from 12.2.0)
+- updated to @applitools/visual-grid-client@15.0.3 (from 15.0.0)
+
+## 1.7.0 - 2020/8/30
+
 - added spec for extracting device name ([Trello](https://trello.com/c/qyf1baqT/464-wdio5-mobile-web-device-name-not-reported-on-dashboard))
+- updated to @applitools/eyes-sdk-core@12.2.0 (from 12.1.2)
+- updated to @applitools/visual-grid-client@15.0.0 (from 14.7.3)
 
 ## 1.6.0 - 2020/8/10