+++ packages/eyes-storybook/CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Unreleased
 
 
+## 3.8.9 - 2020/8/31
+
+- fix exception when getStoryData throws an error ([Trello](https://trello.com/c/FLDMYl2O))
+- updated to @applitools/eyes-sdk-core@12.2.0 (from 12.1.4)
+- updated to @applitools/visual-grid-client@15.0.0 (from 14.7.5)
+
 ## 3.8.8 - 2020/8/13
 
 - updated to @applitools/eyes-sdk-core@12.1.4 (from 12.1.0)