+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -1,7 +1,14 @@
 # Change Log
 
 ## Unreleased
+- updated to @applitools/eyes-sdk-core@12.2.3 (from 12.2.0)
+- updated to @applitools/visual-grid-client@15.0.3 (from 15.0.0)
 
+
+## 5.22.0 - 2020/8/30
+
+- updated to @applitools/eyes-sdk-core@12.2.0 (from 12.1.3)
+- updated to @applitools/visual-grid-client@15.0.0 (from 14.7.4)
 
 ## 5.21.1 - 2020/8/12